backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
pydantic
pydantic-settings
python-multipart
python-dotenv
transformers>=4.35.0
sentence-transformers>=2.2.0
chardet 
torch>=2.0.0
pdfplumber
python-docx
PyPDF2
numpy
accelerate
huggingface-hub>=0.19.0
groq
aiofiles
asyncio
safetensors>=0.4.0
tokenizers>=0.15.0